Add automatic level generation from .txt files
This commit includes a lot.
The most notable changes include the fact that level 4 is auto-generated from a simple .txt file Assets/Resources/LevelData/level4.txt.
This level data must be stored in assets/resources in order to be included in the build: otherwise managing path names when building for different platforms gets very sketchy: working in the unity editor is no indication that it will work in the build.

The format of these level data files is described in Assets/Resources/LevelData/README.md.

The main level-building script LevelBuilder.cs uses prefabs for both nodes and line connections to construct the level.
These both have their own scripts and attributes which can be used to manage events, attribute changes, etc.

The level data is read as a text file, converted to a byte stream, and read back through a stream reader... could use some re-evaluation.