Add attacking line renderers and supporting methods
Right now, a rudimentary animation is thrown together in the Node_Handler
script which shows how to use these methods (defined in the Line Controller script)

The current script works as follows:
When a node is clicked, it picks a random neighbor to attack.
When it reaches the neighbor, it simulates another click on this neighbor.
Each node can only attack once: clicking a node that has already started an attack will not start a new one (this is to avoid infinite recursion)