Fix several bugs, get ready to iplement attack collisions
Such as race conditions for setting an player's start node's shape between different Start() methods

The enemy start node is set to the highest index temporarily, just for development.
Just to simulate an enemy attack before the actual CPU opponent is developed, pressing the
"a" key will cause this enemy start node to attack it's lowest index neighbor.
This is not protected at all, so pressing it when this neighbor is already captured will cause
LineController.StartAttack() to throw a friendly fire error.

There are still bugs with the game controllers' approach to selecting nodes and starting attacks.
For example, a pulse is not stopped with the source node is captured.
(Amusingly, this would allow the previous owner to start an attack on behalf of the new owner)
The order of shapes in transient shape selection behaves strangely as well, especially when changing the potential target